Rename data package to exception which fits better
