+++ src/main/java/duke/command/DukeCommand.java
@@ -1,6 +1,6 @@
 package duke.command;
 
-import duke.data.DukeException;
+import duke.exception.DukeException;
 import duke.storage.Storage;
 import duke.task.TaskList;
 import duke.ui.Ui;
