+++ src/main/java/duke/exception/DukeException.java
@@ -1,4 +1,4 @@
-package duke.data;
+package duke.exception;
 
 /**
  * Signals that an error has occurred during the usage of duke.
