+++ src/main/java/duke/parser/Parser.java
@@ -7,10 +7,10 @@ import duke.command.DukeCommand;
 import duke.command.Exit;
 import duke.command.Find;
 import duke.command.List;
-import duke.data.DukeException;
-import duke.data.InvalidDateTimeException;
-import duke.data.NoToDoDescriptionException;
-import duke.data.UnknownCommandException;
+import duke.exception.DukeException;
+import duke.exception.InvalidDateTimeException;
+import duke.exception.NoToDoDescriptionException;
+import duke.exception.UnknownCommandException;
 import duke.storage.Storage;
 import duke.task.Deadline;
 import duke.task.Event;
