+++ src/test/java/duke/parser/ParserTest.java
@@ -1,6 +1,6 @@
 package duke.parser;
 
-import duke.data.DukeException;
+import duke.exception.DukeException;
 import duke.storage.Storage;
 import duke.task.TaskList;
 import duke.ui.Ui;
